camera and lycoris adjustments
- Camera is now fixed in the X direction.
- Adjusted lycoris distribution.
- Enlarged spirit and lycoris sprites.

+++ Source/Game/Components/WindyComponent.cs
@@ -14,7 +14,7 @@ namespace LifeDeath.Components
         float ax, ay;
         Random rand;
 
-        public WindyComponent(int x, int y)
+        public WindyComponent(float x, float y)
         {
             this.x = x;
             this.y = y;
